Fix segfault on exit when update check is enabled
Before this patch, the script could often crash with segfaults on Debian
12/Python 3.11. It apparently when processing small files.
In that case, there was likely a race condition where Python tried to
terminate the update thread while it was already dead.

It was easy to reproduce with:
```sh
for i in {1..100}; do
    ./MCE.py -skip -exit /lib/firmware/amd-ucode/microcode_amd_fam16h.bin >& /dev/null || echo failed at iteration $i
done
```
GDB showed:
```
Thread 1 "python3" received signal SIGABRT, Aborted.
__pthread_kill_implementation (threadid=<optimized out>, signo=signo@entry=6, no_tid=no_tid@entry=0) at ./nptl/pthread_kill.c:44
44      ./nptl/pthread_kill.c: No such file or directory.
[Thread debugging using libthread_db enabled]
```

Using multiprocessing with daemon=True, there are no more crashes.
Processes are probably easier to kill properly. The result of the
update check is now shared using a queue.